All recipies displaying using bootstrap cards

+++ recipies/models.py
@@ -3,7 +3,7 @@ from django.db import models
 
 class Recipe(models.Model):
     recipe_name = models.CharField(max_length=255)
-    prepare_time = models.FloatField()
+    prepare_time = models.IntegerField()
     portion = models.IntegerField()
     image_url = models.CharField(max_length=2083)
     recipe_url = models.CharField(max_length=2083, default="")
